Misc: Add editorconfig, prettier and sync lint rules
The lint rules came from GNOME Shell's repo but have been outdated for a
while. So update them now (and transform them to the flat file format).
Only change is that I will always require curly braces as I am not a
fan of prettier inlining expressions with something like if clauses.