Misc: Add .git-blame-ignore-revs file
The previous commit did a massive amount of changes because of the
introduction of new eslint rules and prettier. To keep git blame clean
a .git-blame-ignore-revs file that can be used to ignore that commit.
Local git users need to use that file with the `--ignore-revs-file` flag
(or add it to their local git config to always use that file). Github
will automatically use that file when using their online blame.